+++ app/Http/Controllers/UsersController.php
@@ -33,6 +33,14 @@ class UsersController extends Controller
         return 0;
     }
 
+    public static function isUserSuperAdmin()
+    {
+        if (User::where('masp', TokenController::$payload->number)->where('isSuperAdmin', 1)->first()) {
+            return 1;
+        }
+        return 0;
+    }
+
     public static function isAdminView()
     {
         return Session::get('admin');
@@ -50,14 +58,15 @@ class UsersController extends Controller
 
     public function index()
     {
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         $users = User::orderBy('created_at', 'desc')->get();
-        //dd($users);
-        return view('admin.admin_panel', ['users' => $users]);
+        return view('admin.index', ['users' => $users]);
     }
 
     public function create()
     {
-        return view('admin.admins_create');
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
+        return view('admin.create');
     }
 
     public function store(UserCreateRequest $request)
@@ -73,42 +82,46 @@ class UsersController extends Controller
         else if ($users->where('masp', $request->masp)->first()) {
             $user = $users->where('masp', $request->masp)->first();
             $user->restore();
-            return redirect(route('admins.index'))->with('status', 'Administrador ' . $user->masp . '  inserido com sucesso!');
+            return redirect(route('admin.index'))->with('status', 'Administrador ' . $user->masp . '  inserido com sucesso!');
 
         }
         $user = new User(request(['masp']));
         $user->save();
-        return redirect(route('admins.index'))->with('status', 'Administrador ' . $user->masp . '  inserido com sucesso!');
+        return redirect(route('admin.index'))->with('status', 'Administrador ' . $user->masp . '  inserido com sucesso!');
     }
 
     public function show(User $user)
     {
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         return view('users.show', ['user' => $user]);
     }
 
     public function edit(User $user)
     {
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         return view('users.edit', compact('users'));
     }
 
     public function update(UserCreateRequest $request, User $user)
     {
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         $request->validated();
-        return redirect(route('admins.index'))->with('status', 'Administrador ' . $user->masp . '  já existe e foi atualizado');
+        return redirect(route('admin.index'))->with('status', 'Administrador ' . $user->masp . '  já existe e foi atualizado');
     }
 
     public function destroy(User $user)
     {
-
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         if ($user->masp == '1729862' || $user->masp == '1292598')
-            return redirect(route('admins.index'))->with('status', 'Não é possível excluir o ' . $user->masp);
+            return redirect(route('admin.index'))->with('status', 'Não é possível excluir o ' . $user->masp);
 
         $user_masp = $user->masp;
         $user->delete();
-        return redirect(route('admins.index'))->with('status', 'Administrador ' . $user_masp . ' apagado com sucesso');
+        return redirect(route('admin.index'))->with('status', 'Administrador ' . $user_masp . ' apagado com sucesso');
     }
 
     public function getAdminUsers() {
+        if(!$this->isUserSuperAdmin())  return redirect(route('documents.index'));
         return view('admin_panel');
     }
